Correction de error de compilacion. Se adapta arcrandom a nueva versión de swift

+++ Hamburguesas/Hamburguesas/ColoresStruct.swift
@@ -29,7 +29,7 @@ struct Colores
     
     
     func RegresaColorAleatorio()->UIColor{
-    let posicion=Int(arc4random()) % colores.count
+    let posicion=Int(arc4random()%UInt32(colores.count))
     return colores[posicion]
 
     }
